requirements/base.txt
celery==4.1.0
django==2.0.3
django-environ==0.4.4
boto3==1.6.11
django-model-utils==3.1.1
pika==0.11.2
psycopg2cffi==2.7.7
python-dateutil==2.7.0
djangorestframework==3.7.7
gunicorn==19.7.1
django-polymorphic==2.0.2
django-rest-polymorphic==0.1.4
